+++ tailwind.config.js
@@ -1,14 +1,22 @@
 /** @type {import('tailwindcss').Config} */
 import colors from 'tailwindcss/colors';
+import defaultTheme from 'tailwindcss/defaultTheme';
 
 export default {
   content: ['./src/**/*.{html,js,svelte,ts}'],
   theme: {
     extend: {
       colors: {
-        primary: colors.violet,
+        primary: colors.indigo,
         neutral: colors.slate,
+        success: colors.green,
+        info: colors.yellow,
+        alert: colors.red,
+        alt: colors.teal,
       },
+    },
+    fontFamily: {
+      sans: ['Inter', ...defaultTheme.fontFamily.sans],
     },
   },
   plugins: [],